+++ spec/jobs/create_event_notification_job_spec.rb
@@ -1,9 +1,24 @@
 RSpec.describe CreateEventNotificationJob, type: :job do
-  let!(:event) { create :event }
+  context 'for guild event' do
+    let!(:guild) { create :guild }
+    let!(:event) { create :event, eventable: guild }
 
-  it 'executes Notificators::CreateEventNotificator.call' do
-    expect(Notificators::CreateEventNotificator).to receive(:call).and_call_original
+    it 'executes Notificators::CreateEventNotificator.call' do
+      expect(Notificators::CreateEventNotificator).to receive(:call).and_call_original
 
-    described_class.perform_now(event_id: event.id)
+      described_class.perform_now(event_id: event.id)
+    end
+  end
+
+  context 'for guild static event' do
+    let!(:guild) { create :guild }
+    let!(:static) { create :static, staticable: guild }
+    let!(:event) { create :event, eventable: static }
+
+    it 'executes Notificators::CreateEventNotificator.call' do
+      expect(Notificators::CreateEventForGuildStaticNotificator).to receive(:call).and_call_original
+
+      described_class.perform_now(event_id: event.id)
+    end
   end
 end