+++ spec/services/notificators/create_event_for_guild_static_notificator_spec.rb
@@ -1,13 +1,14 @@
-RSpec.describe Notificators::CreateEventNotificator, type: :service do
+RSpec.describe Notificators::CreateEventForGuildStaticNotificator, type: :service do
   let!(:guild) { create :guild }
-  let!(:guild_event) { create :event, eventable: guild }
-  let!(:notification) { create :notification, event: 'guild_event_creation' }
+  let!(:static) { create :static, staticable: guild }
+  let!(:guild_event) { create :event, eventable: static }
+  let!(:notification) { create :notification, event: 'event_creation_for_guild_static' }
   let!(:delivery) { create :delivery, deliveriable: guild, notification: notification }
   let!(:delivery_param) { create :delivery_param, delivery: delivery }
   let(:service) { described_class.new }
 
   describe 'self.call' do
-    it 'executes call for Notificators::CreateEventNotificator object' do
+    it 'executes call for Notificators::CreateEventForGuildStaticNotificator object' do
       expect_any_instance_of(described_class).to receive(:call).and_call_original
 
       described_class.call(event_id: guild_event.id)