+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- notification for event_creation_for_guild_static
+
+### Fixed
+- bug with static show
+
 ## [0.9.1] - 2019-11-30
 ### Added
 - model GroupRole